docusaurus-book/backend-for-rag chatbot/pyproject.toml
[project]
name = "rag-backend"
version = "0.1.0"
description = "RAG backend for Physical AI Humanoid Robotics documentation, using Cohere embeddings and Qdrant vector database for semantic search and question answering"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "cohere>=5.20.0",
    "qdrant-client>=1.16.2",
    "requests>=2.32.5",
    "trafilatura>=2.0.0",
    "python-dotenv>=1.0.0",
    "openai-agents>=0.1.0",
    "openai>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pydantic>=2.0.0",
]
